sanitize

+++ src/ctrip_swap_repl.c
@@ -226,7 +226,7 @@ static void processFinishedReplCommands() {
 			if (applied) {
                 gno_t gno = 0;
                 char *uuid = NULL;
-                int uuid_len = 0;
+                size_t uuid_len = 0;
                 if (gtid_repr) {
                     sds repr = gtid_repr->ptr;
                     uuid = uuidGnoDecode(repr,sdslen(repr),&gno,&uuid_len);
